requirements.txt
Django==4.2.4
aiogram-3.0.0rc2
environs-9.5.0
